Move integration tests to tests/integration
